Renaming parser to data module
